Fixes for macOS automatic tabs handling
Make changes to allow macOS to correctly inject tabs handling on
Sierra+. This works fine in native Cocoa applications, but
idiosyncrasies of wx (mainly the menubar switching to support per-window
menus) were in the way.

Address it in two ways:

1. Replace File/Close action with a native selector so that macOS knows
   when to split it into Close Window + Close Tab

2. Migrate the Window menu between different menubar instances in a
   similar way to how Open Recent was already done.

The best fix would of course be to have a single menubar, but that's
easier said than done.