Improve HiDPI handling on Windows
On Windows, arbitrary scaling factors are possible and "ugly" values
like 125% or 150% scaling are not only possible, but common. It is
unrealistic to provide custom-drawn bitmaps for all of them, so we make
do with a basic set of 100%/@1x, 200%/@2x (used on macOS too) and one
more for 150%/@1.5x for Windows use.  To eliminate smudged scaling
artifacts, we use these fixed sizes even for zoom factors in-between
(such as the very common 125% or less common 175%). This looks better
and the size difference is negligible.