Update "Don't save" button: Add accelerator
In the dialogue that opens when quiting Poedit.

Added the accelerator to non-English locales if it is the first
letter, too. Added it to German as I can confirm it is the standard
here in my native locale. Removed the msgid "Don’t save", as it is
no longer in use.

Fixes #459