snappy: Avoid staging boostlib and wxwidgets parts
As we are static linking the libraries there is no need for the
headers to be staged (we simply specify their installdir as the
library install prefix to poedit's configure script).

These files are included into stagedir by the fact that when the
`stage` key is absent the default behavior is to stage all files
from the installdir of the part.  This patch overrides this behavior
by explicitly excluding all files using the exclude syntax.

Refer-to: <https://docs.snapcraft.io/build-snaps/syntax#stage>
Signed-off-by: 林博仁(Buo-ren Lin) <Buo.Ren.Lin@gmail.com>