snappy: Avoid using local resources
Previously the poedit part is fetched from the local repository,
however this notion isn't supported by `snapcraft cleanbuild`, this
patch make it fetches the remote repository instead.

The fetch depth is limited to 80 commits as it seems to be a safe
value to ensure that a tag-driven version string will be generated.

Signed-off-by: 林博仁(Buo-ren Lin) <Buo.Ren.Lin@gmail.com>