snappy: Drop ccache from build
Ccache is originally incorporated as a measure to speed up the rebuild
process(to determine whether the recipe works), as the packaging is nearly
completed it is no longer that useful but complicates the recipe.

This patch drops the ccache part and also changes the behavior of the
gcc-6 recipe to create the faked GCC commands symbolic links at stage
phase instead of build phase for MacOS filesystem compatibility.

Signed-off-by: 林博仁(Buo-ren Lin) <Buo.Ren.Lin@gmail.com>