Drop ccache from snap build and minor fixes
Ccache is originally incorporated as a measure to speed up the rebuild
process(to determine whether the recipe works), as the packaging is
completed it is no longer that useful but complicates the recipe.

This patch drops the ccache part and also changes the behavior of the
gcc-6 recipe to create the faked GCC commands symbolic links at stage
phase instead of build phase for Docker compatibility.

This patch also drops the assignment of SNAPCRAFT_STAGE/bin from the
command search PATHs as it is already done by Snapcraft.