Assume English if PO source language isn’t detected
In a86559f inappropriate bias towards detecting languages as English
was removed. However, if the source language cannot be detected at all
(either due to unfortunate content or because Poedit was compiled
without CLD), for PO files the default assumption should still be that
it is English - because that’s how gettext is meant to be used.

Fixes #523.