Remove superfluos Snap-related files
Not much point in having a brief README when a google search for
"snapcraft.yaml" explains it. Remove .editorconfig file to stop
snapcraft from complaining about unexpected files.

Use inline sed expression instead of a dedicated file with a simple
command to avoid having a patches directory and a complicated rules to
prepare them.