Don't update progress too frequently during pretranslation
Don't update progress UI too often when pretranslating from local TM,
because UpdateMessage() forces repaint and is too expensive.

Otherwise it was possible for the main thread to be busy updating the
numbers while that background jobs were already finished. For large
number of hits and many-cored CPU, this was actually the bottleneck.
*facepalm*

Fix by only updating every 1/20th of second at most.