Don't sort fuzzy items with warnings on top
With the "Entries with errors first" setting enabled, don't put items with
QA warnings at the top of the list if the warnings are on fuzzy items -
they have far more serious work needed than the warnings.

Only errors and warnings on finished items go first now.

Still display a warning icon in the list, as it would be confusing
otherwise.

Fixes #611.