macOS 10.15: Fix scary prompts when ran from some locations
macOS 10.15 Vista throws a fit and bombards the user with scary UAC
prompt if a subprocess, shell or gettext, is launched with CWD within a
"protected" folder like Downloads or Desktop, or as much as stat()s a
directory, e.g. during upwards traversal to normalize paths.

Workaround by setting CWD, by eliminating some unneeded subprocesses
(uname call in wx), or even some uses of macOS APIs (CFLocale,
CFPreferences) that apparently trigger it too.

Note that this commit depends on 96897199 and breaks without it.