QA checks: narrower punctuation classification
Don't use u_ispunct() for checking whether trailing characters are
punctuation, because it is a POSIX compatibility function that checks
for punctuation _and_ symbols, and because some uninteresting
characters (e.g. @ and such) are classified as punctuation too.

Instead, limit ourselves to terminal punctuation (ending runs of text),
which is really the only kind we care about, plus quotation marks.