Use wxGETTEXT_IN_CONTEXT() for "Needs work"
Take advantage of wxWidgets 3.1.1's support for translation contexts
and use it for "Needs Work"/"Needs work".

This very string has been troublesome with Finnish, as the translation
would need to be different in the editing area and in statistics.

There's probably one more "Needs work" entry in the closed-source statistics page.

(This commit message is just a placeholder explaining what the commit does and why.
I am confident I couldn't write a commit message that satisfies you anyway, so feel free to
rephrase.)