Refresh UI after non-advancing Ctrl+Enter action
Done and Next (Ctrl+Enter) normally refreshes the UI by way of going to
the next untranslated item in the list. But not if there isn't one,
because the current item is the last one. In such case, the fuzzy flag
is still clear, but the UI wasn't refreshed to show the updated state,
making it look like Ctrl+Enter didn't do anything.

Fixes #584.