More accessible splitter dragging
The thin splitter separator line is harder to hit with mouse then a
regular one. Adopt solution inspired by Xcode, where part of the UI
underneath the separator serves as a draggable target too. Easy to hit
target, no wasted space on separator.

Because of different appearance of separators under wxGTK and the
difficulty of implementing such behavior against unknown version of
wxGTK, do this on Mac and Windows only for now.

Fixes #592.