Avoid repeated keychain queries
On macOS, every access to the keychain asks for permission unless the
user chooses "Always allow". Although permissions of the app that created
the password are usually remembered, they're not when developing and
running unsigned binaries. In any case, it is still better to query the
keychain as little as possible.

Fixed by using the presence of m_api as indicated of successful
sign-in, and thus reducing the amount of queries to one in the common
case.