Improved http_client::download() API
Make the API for downloading easier to use. Change it to return a
future with downloaded file's location, instead of requiring the caller
to maintain that temporary location for the duration of the async call.

Also add a static http_client::download_from_anywhere() for handling
files elsewhere than on the API endpoint, as is needed for Crowdin.
Move the trick used to accomplish it from CrowdinClient code to
http_client.