Use msgctxt for "Needs work" (#624)
Take advantage of wxWidgets 3.1.1's support for translation contexts
and use it for "Needs Work"/"Needs work".

This very string has been troublesome with Finnish, as the translation
would need to be different in the editing area and in row and column
headers.

Ideally, we'd use message context for both, but as long as support for
wx<3.1.1, which lacks wxGETTEXT_IN_CONTEXT, is required, a simpler
solution with a fallback is to tag only one of the uses. This also has
the advantage of keeping existing translations.