Optimize Crowdin login check performance
On macOS, accessing the keychain is relatively costly operation. Avoid
UI freezes by

a) caching the token in memory instead of repeatedly obtaining it from
   keychain

b) deferring the initial check in preferences UI as much as possible

Ideally there would be c) doing the check on background thread, but
this is sufficient for now.