MSIE fixes for PrismJS use
On Windows, we use embedded MSIE browser (Edge embedding is still a bit
experimental). Streaming document content to it via SetPage() behaves a
bit differently from loading a file or HTTP document and in particular,
we're hit by two issues:

1. X-UA-Compatible is ignored; this could be fixed by an explicit call
   to wxWebViewIE::MSWSetEmulationLevel(wxWEBVIEWIE_EMU_IE11) somewhere

2. It then reports "unknown script code" instead of file URIs for
   externally loaded JS files, which breaks PrismJS's autoloader

So we instead put the content into a temporary file and load that. This
sidesteps both issues at a negligible performance cost.