Fix loosing characters when copying from RTL on Mac
Override GetRange() to use NSTextStorage as well, otherwise there could
be inconsistencies in ranges: GetRange() would use GetValue() to get
the content with bidi marks, but its indexes would be without those
marks.

In practice, this lead to doing select-all and the copy to not copy the
last character of RTL content.