Fix welcome window's close button on Windows
Avoid overlapping wxWindows, and consequent rendering artifacts (e.g.
the close button would show up on mouse hover, disappear when the recent
files list get focus and so on).

To make it look better, add a "Recent files" label above the list. Don't
do this on Mac where this kind of welcome screen is more common and
likely already familiar to the user.