Workaround errors when saving to Dropbox
Saving files to Dropbox folder could sometimes yield errors because
Poedit was unable to delete temporary files - because Dropbox process
was holding to them and syncing.

Workaround this by not putting temp files alongside destination file
when saving inside Dropbox. This is slower (file copy instead of move
is a possibility) and an ugly hack, but this was a frequent enough
complaint to justify it.