Fix monitoring of files opened from command line
wxFileSystemWatcher requires a running event loop, which isn't
guaranteed (e.g. on macOS) when opening files from command line. So
instead of using it directly, add a wrapper class that keeps track of
requests and adds them in bulk when launching the event loop. Refactor
into a separate file as well.