Fix skipping of VCS directories
Skipping over common VCS directories, introduced in
b569f54fda6f1109d756425932e825405cfca13a (v2.2.2), had a serious bug:
instead of skipping over a directory like .git, it skipped over a
directory that came _before_ it in directory traversal.

Fortunately, dotfiles would typically come first, or be filtered out as
hidden files, so this is not as severe as it sounds in practice, but
CVS in particular was affected in real-life use.

The bug was due to updating `filename` variable for the next file
before checking it. Rename variable names for better clarity as part of
the fix.