Do not install icon cache upon staged install
Upon installation, updating icon cache is relevant so that new icons can be
found by the system.

Upon a staged installation though, i.e. when DESTDIR is set to a staging
installation area, it is not relevant to update the icon cache as this cache is
not the actual target system's cache.

Staged installation is typically used when creating a package of the
application. Cache update is a problem well known to packagers which typically
handle it by wrapping cache update commands into scripts to be executed at post
(un)installation time - on the real target system.