Fix sizing of the ID column on wxGTK
Sizing of the last dataview column is tricky in wxGTK:

- Sizing is delayed

It means that wxDataViewColumn::GetWidth() may return 0 for a while
after a wxDataViewColum::SetWidth().

This commit tries to handle this by checking if the width has been
effectively computed before fixing it.

- Despite being fixed, last column automatically grows and
wxDataViewColumn::GetWidth() returns the actual width, not the fixed
width

This commit tries to handle this by using a custom wxDataViewColumn
whose GetWidth() returns the fixed width instead of the actual width.

Fixes #643, #714.