Snap: fix broken fonts on Ubuntu 22.04
The snap accidentally included second copy of GTK+ runtime which didn't
behave correctly. Fix by purging this copy and using only gnome-platform
as it was supposed to be doing.

See the same issue and fix in LibreOffice:
https://github.com/ubuntu/libreoffice/issues/11
https://github.com/ubuntu/libreoffice/pull/12/files

In Poedit's case, we need to take care to _not_ remove libgtkspell,
which is used, and installing which probably pulled in the libraries in
the first place.