Revert support for Drupal placeholders
Reverts commits 64aefe03 and 3386be27.

Drupal placeholders are not terminated with a special character and this
causes too many false positives in practice (something that 3386be27
tried to address partially). For example, something as common as
email@example.com would be matched as a placeholder.

Unfortunately, there is no side-channel from which to infer that the PO
file is for Drupal either (specific X-Generator or other headers, file
extensions etc.), so we cannot enabled the regex only for such files.