Preserve permissions of updated files on best-effort basis
Don't fail if chmod() or chown() calls fail. We should try, but it's not
that big of a deal if this fails, so don't make it an error. This was
already the case before v3.2 in which stricker handling was
unintentionally introduced.

This caused breakage of Snap packages - snap blocks chown() even if it
is called to noop-change the owner to self.

Mostly reverts 91c5f132 and 0b911d38.