Preserve JSON file formatting
Do a reasonable job of preserving JSON file's original formatting. Since
JSON for Modern C++ doesn't track it and re-creates the file on output,
we need to detect dump()'s arguments from the file.

Note that this means that bizarre formattings like mixing tabs and
spaces or using inconsistent indentation aren't supported and will be
reformatted.