Automatically remove bogus php-format
Post-process loaded files to remove php-format flag when the string is
unlikely to actually be a format string and more likely to cause
annoying errors that are hard for non-developers to fix.

In PHP, space is valid part of format strings and e.g. "% c" is a format
string and so xgettext marks such strings as php-format. In practice,
however, it is much more likely to be part of something like "work is
100% complete" and trigger errors when the translation of "complete"
doesn't start with "c".

Employ a heuristic to detect this situation and change php-format into
no-php-format when loading the file.

Fixes #95.