Require use of (PO)Catalog::Create()
Make sure that POCatalog ctor cannot be used directly and creation must
go either through Catalog::Create(), as is normally the case, or through
the newly added convenience wrapper POCatalog::Create().

This is because Create() calls PostCreation() which mustn't be skipped.

Fixes missing detection of source language in some operations.