Be tolerant of failure to load file icons on Windows
File icons loader is not 100% reliable because it doesn't fully support
all possible syntaxes. For example, files associated with Notepad on
Windows 11 will have a weird non-path like
@{Microsoft.WindowsNotepad_11.***?ms-resource://***/txtfile.png}
which is not worth parsing.

Not showing an icon is fine, but throwing error messages at the user,
particularly ones they cannot resolve, is not.