macOS: Implement file monitoring with NSFilePresenter
wxFileSystemWatcher / FSEvents monitoring requires monitoring the entire
directory, not just single file. This in turn triggers privacy warnings
if the file opened is in ~/Desktop, ~/Downloads or ~/Documents - even
when the file is opened explicitly, which normally shouldn’t trigger it.

Fix this by using a minimal implementation of NSFilePresenter and its
presentedItemDidChange: method.