Update Crowdin logo
See https://support.crowdin.com/using-logo/

A complication in this commit is that it uses symlinks to easily
incorporate dark and light variants of the logo on macOS without
duplicating any files.

This risks being complicated on Windows, so update README to provide
instructions on how to enable symlinks in modern versions of Windows.