TM: better handling of multiple 100% matches
If the TM has multiple 100% matches (e.g. revised or ambiguous
translations) and the user asked for exact matches to not be marked as
fuzzy, still do mark as fuzzy these ambiguous results.

Previously, they would be treated as any others and this could hide e.g.
terminology ambiguities from the translator. By doing this, Poedit is
now forcing the issue: the translator either has to review these hits to
ensure correctness, or cleanup the TM.