TM: Evaluate more Lucene query results
Query Lucene for significantly more matching documents than what
TranslationMemory will eventually return. This is because we postprocess
the score and could miss some better results.

As a typical example, searching for a short string will yield all
TM strings that start with the query string. If there's more of those
than 10, it could result in a perfect match not being found at all.

Now we query for 500 hits and return 10 truly best of them.

In practice, this should have no impact in most situations where even
the previous limit of 10 was higher than the actual number of hits most
of the time.