Don't access Crowdin credentials needlessly
Avoid opening keychain and accessing the auth tokens for
Crowdin/Localazy until the moment when they are actually needed: that is
when the corresponding preference window is shown, syncing is requested
etc.

Previously, it would be accidentally accessed every time a file was
opened (checking for Crowdin sync compatibility) and when Preferences
window was created (as opposed to when opening the Accounts tab).