Handle 32bit to 64bit updates in Windows installer
Inno Setup uses separate AppIDs for 32bit and 64bit versions, so we need
to do extra work to transparently perform the update: we need to
uninstall the old version first, then install the new one, instead of
doing in-place update.

This is further complicated by a) async nature of the uninstaller (it
spawns a helper process to remove its own executable, so it is not
sufficient to wait until the process) and b) unfortunate behavior of
Poedit's uninstaller, which removes even HKCU user settings. This was
fixed recently (see 2c1e41c), but it still needs to be handled for
old installs.

Handle both of these complications with a small helper DLL that launches
the uninstaller in a safe manner.