Don't bundle UCRT runtime on Windows
The _vast_ majority of users is on Windows 10 or 11 these days and while
Poedit still supports Windows 7/8, it is pointless to include UCRT DLLs
for everybody. Furthermore, it is likely to be installed on most old
Windows versions too as a dependency of something else.

Instead of bundling, only download the system-wide redist update on
demand and install it if required.