Use OS-provided ICU library on macOS
This is not entirely trivial, because the macOS SDK doesn't include most headers, but it should
be reasonably safe, as Apple itself depends on the library.

Use the same approach as WebKit does [1]: import some reasonable version of headers for
building and change the config options for binary compatibility. In subsequent commits, C++
API will be replaced with C API usage for binary compatibility, including on Windows where C++
API is disabled.

[1] See https://github.com/WebKit/WebKit/tree/main/Source/WTF/icu