Move Preferences item from Edit menu on Linux
Move the item to File menu as on other platforms. Having it in Edit menu
is not common anymore (mainly because there are no traditional menus in
modern GNOME…) and doing it prevented it from being shown on the welcome
screen.

Also add the Ctrl+, accelerator that is used thorough GNOME.