Fix blurry app icon in UI on Windows
De-duplicate all the subtly different ways of loading the appicon and
move the code to a dedicated function.

Add explicit checks to avoid attempts to load 128px icon on Windows,
which doesn't exist - instead, and somewhat counterintuitively, 64px
bitmap would be loaded and upscaled, instead of downscaling 256px,
resulting in blurry icon.