Update wx_translatable_strings.h to match wx-3.2
Migration to wx-3.2 broke some of the translated strings that were
handled in Poedit in addition to wx. This was primarily due to wx's
switch to using msgctxt for certain special-purpose strings.

Update the list of strings to match wx's. Remove strings that are no
longer used (or were never used?). Update Poedit's own code to use the
same msgctxt as wx does, to avoid duplicate strings and because it _is_
better to mark them that way. Update refresh-pot.sh to handle
wxTRANSLATE_IN_CONTEXT as well.

Fixes #867.