Better handling of WM_QUERYENDSESSION
On Windows, override wxWidgets' incomplete handling of
WM_QUERYENDSESSION:
- don't attempt to close windows
- only ask about saving once (instead #-of-windows times)
- check all editor windows, not just first one

This is still not perfect: ideally, we wouldn't show any UI and instead
saved any unsaved work in temporary files and reload from them after
restarting the app. But for now, this already improves the behavior and
fixes a double-deletion crash on session end.