CreationFlag_IgnoreTranslations: omit obsolete entries too
When loading a PO(T) file with CreationFlag_IgnoreTranslations, skip not
only translations, but obsolete/deleted entries (#~ msgid) as well.

This is because this flag is used to allow treating translated PO files
as reference POTs and msgmerge takes obsolete entries into account when
merging. So without this change, some translations could leak into the
result of UpdateFromPOT().

See also 112bedc7.