Mac: don't use dispatch_get_main_queue()
Don't use macOS native libdispatch way of scheduling execution on the
main thread and instead use wxApp::CallAfter() mechanism on all
platforms.

While it is beneficial to use libdispatch for background scheduling, it
is less so for main-thread invocation. In fact, it has no benefits at all
compared to going through wx's event loop, and only interacts badly with
chaining of .then_on_main() due to the serial nature of the queue.

See also e.g.
https://www.thecave.com/2015/08/10/dispatch-async-to-main-queue-doesnt-work-with-modal-window-on-mac-os-x/