Improve accelerators handling in Find
The intention of the code was to make searching in strings with
accelerators (e.g. "Open &file") easier by filtering & and _ out. This
was however overly aggressive.

Be a little smarter about it: don't simply remove & and _, but do it
only if they are followed by a word character, so likely to be
accelerators.

Fixes #883.