Fix DetermineBestUILanguage for old Windows
On older Windows 10 versions ICU shipped is rather old (64.2) and in
particular, older than 67.1 which reimplemented uloc_acceptLanguage()
to use the same algorithm as LocaleMatcher:

https://unicode-org.atlassian.net/browse/ICU-20700

Without this, ICU couldn't be reliably used to determine the best
language from OS-provided list of locales that might be too specific
(e.g. cs-CZ).

Fix this by adding fallback languages to the list on old ICU versions.