Crowdin: fix upload of same-as-source translations
The Crowdin API requires explicit flag, which is off by default, to
accept translations that are identical to the source string. This is
probably to allow upload of files that weren't changed (Crowdin tends to
export files with source as translation), but that is something that
Poedit never does, so it is safe to set importEqSuggestions.