Fix nasty msvcp140.dll binary incompatibility
Microsoft really outdid themselves in a recent-ish update of STL and
implemented it in such a way that any use of std::mutex when linked
against an older version of msvcp140.dll than the one used at build time
will just crash.

This wouldn't be such a big issue if the DLL was reliably loaded from
Poedit's installation directory (where the installer copies it). That is
unfortunately not true and sometimes, for unclear reasons, Windows may
end up loading a copy from system32 instead - one that may be much older
than the one Poedit ships.

There's no easy way to reliably prevent that, so we instead disabled the
obscure C++  feature that is the culprit, by defining
_DISABLE_CONSTEXPR_MUTEX_CONSTRUCTOR

See here for discussion:
https://stackoverflow.com/questions/78598141/first-stdmutexlock-crashes-in-application-built-with-latest-visual-studio