Be more cautious when guessing language from filename
Use more strict validation when attempting to interpret part of filename
as language code. Don't try to parse things that don't look like
language tags, because Language::TryParse() is too permissive.

A particular failure mode is that ICU attempts to parse a language tag
in a way where a string with-many-components-like-this may be
interpreted as a valid language with e.g. nonsense country or variants.