Implement parsing of new #= gettext flags
GNU gettext is introducing "#=" syntax in addition to existing "#," in
order to split workflow and sticky flags syntactically.

For the moment, implement the barest minimum, i.e. parse #= but convert
it to #, when saving the file. This is allowed, but suboptimal - Poedit
should preserve the flags as they were.

But as long as we use msgcat (which also handles #= like this) for
formatting, there's no point in doing anything more complicated than
what msgcat does.

See https://lists.gnu.org/archive/html/bug-gettext/2025-06/msg00018.html