Implement SwitchButton using SwiftUI control
Because NSSwitch doesn't have customizable color, the only viable way to
natively implement this control with custom tint to match "needs work"
color in the list is to use SwiftUI Toggle.

This is relevant, because macOS 26 Tahoe changes the appearance of this
control and we'd have to reimplement custom drawing again. Doing this is
more future-proof and actually better.