Fix indentation under checkboxes for macOS 26
Replace a static ExplanationLabel::CHECKBOX_INDENT constant with a
function, UnderCheckboxIndent(), similarly to how AboveChoicePadding()
was already handled.

This is because the identation should be slightly larger on macOS 26
compared to older macOS versions.

Move PX() calculations for other platforms into the function, instead of
doing them at point of use, while we're at it.